requirements.txt
click==7.1.2
CouchDB==1.2
exif==1.2.1
Faker==6.6.3
Flask==1.1.2
future==0.18.2
guizero==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
opencv-python==4.5.1.48
piexif==1.1.3
Pillow==8.1.2
pkg-resources==0.0.0
pkinter==1.37.2
plum-py==0.3.1
python-dateutil==2.8.1
six==1.15.0
text-unidecode==1.3
tkintertable==1.3.2
Werkzeug==1.0.1
